WoT restructure: nest rangeQuery/ under table/ (table-family is its only consumer)
