+++ handsontable/src/3rdparty/walkontable/src/table/baseTable.ts
@@ -1,7 +1,7 @@
 import type { WalkontableInstance } from '../types';
 import type { EngineContext } from '../wire';
 import type Settings from '../settings';
-import type { RowRangeQuery, ColumnRangeQuery } from '../rangeQuery/renderedRange';
+import type { RowRangeQuery, ColumnRangeQuery } from './rangeQuery/renderedRange';
 import { cellAccess, type CellAccess } from './cellAccess';
 import { domScaffold, type DomScaffold } from './domScaffold';
 import {
