+++ handsontable/src/3rdparty/walkontable/src/table/rangeQuery/renderedRange.ts
